refactor vyper to work with context command.
